Decompose tab-drag.js into three terminal-* modules
The 962-LOC tab-drag.js bundled five disparate concerns. The split:

- terminal-pointer.js — chip drag, splitter resize, pane-resize drag.
- terminal-lifecycle.js — tab create/close/rename + restore-on-reload.
- terminal-shortcuts.js — keybindings + screenshot-paste.

Each module exposes its own initSideEffects() that dashboard-main.js
registers in sequence. Cross-module references live inside function
bodies so the live-binding cycle stays TDZ-safe (rules carried over
from notes/01-p07-tab-drag-split.md §D2). The duplicate `var
_termDrag` from the original (chip-drag and pane-drag both reused
the binding) becomes two distinct names, one per module.
